Update nfs sudo configuration for Fedora
Copied the configuration to my own machine and noticed I was still getting prompted for my password. Looking at the command, I realize that sed is called on /tmp/exports, not /etc. Quick change, and everything is automatic as expected!

Updating the documentation so others don't have the same hassle. :)